@@ -9,9 +9,10 @@
     "bootstrap": "^3.4.1",
     "immutability-helper": "^2.1.1",
     "keycode-js": "^0.0.4",
-    "react": "^15.4.2",
-    "react-dom": "^15.4.2",
-    "recompose": "^0.23.5"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "recompose": "^0.23.5",
+    "sweetalert2": "^11.15.0"
   },
   "scripts": {
     "start": "react-scripts start",
