@@ -1,5 +1,5 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';  // Make sure to import from 'react-dom/client'
 import App from './components/wrappers/App';
 
 // Add bootstrap
@@ -8,7 +8,10 @@ import 'bootstrap/dist/css/bootstrap.css';
 // Add our style
 import './assets/style/index.css';
 
-ReactDOM.render(
-    <App/>,
-    document.getElementById('root')
+// Create the root element and render the app
+const root = ReactDOM.createRoot(document.getElementById('root'));  // Create a root
+root.render(
+    <React.StrictMode>
+        <App />
+    </React.StrictMode>
 );
